gcloud_functions/requirements.txt
beautifulsoup4==4.8.1
Pillow==6.2.1
lxml==4.4.1
requests==2.22.0
